[Proposal] Solution 2: PlaybackObserver utils moved to core/common
Root issue
----------

While working on code refactoring for better taking into account the new
potentially multithread nature of the RxPlayer code (with some files only
running in a WebWorker environment and others only running in a main
thread environment), we recently refactored our file hierarchy (#1365) to
better reflect that new situation. The idea is to make RxPlayer
developpers more aware of what code is intended to run where.

In that work, we had a remaining issue concerning the `PlaybackObserver`.
This is the part of the code that is monitoring and advertising playback
conditions to the rest of the RxPlayer code. Most core RxPlayer modules
rely on it, in both main thread and WebWorker environments.

The root issue is that the `PlaybackObserver` is a class and thus cannot
be easily transmitted in-between environments (as the main thread and
WebWorker only exchanges between one another through `postMessage`
calls, which follows some rules preventing from doing that).

As such, and only on a multithreaded scenario, the RxPlayer is serializing
in some way the constructed source PlaybackObserver in the main thread to
reconstruct one (the `WorkerPlaybackObserver`) on the worker.

Because a whole complex PlaybackObserver-compatible structure has to both
be constructed in the main thread and in the worker (yet only the main
thread can act as the "true" source one, because the media element can
only be accessed in main thread), we were asking ourselves where should
we put the common utils (required by both the main thread and worker)
needed to construct one (like the `ObservationPosition` class and the
`generateReadOnlyObserver` util).

Solution 1
----------

This is the second solution (after #1366), which only moves utils common
to both the WebWorker's PlaybackObserver and the main thread's
PlaybackObserver into the `src/core/main/common/` directory - just like
`ContentInitializer` utils in monothreaded mode.

In a way, this solution repeats the same solution found for
`ContentInitializer` utils - which are in a different place when ran in
monothreaded mode (where the utils is imported by the
`src/main_thread/init` logic) or ran in multithreaded mode (where the
utils imported by the `src/core/main/worker` logic).

Result
------

I will admit that I'm less a fan of this solution than #1366, as the
`src/core/main/common` feels like a hack in any scenario it is used.

There is also the fact that types exported through that path need to be
re-exported through a more accessible path (usually
src/main_thread/types`) to simplify import paths in the project.

Thoughts?
