+++ src/main_thread/api/index.ts
@@ -18,10 +18,10 @@ import PlaybackObserver from "./playback_observer";
 import Player from "./public_api";
 export { PlaybackObserver };
 export {
-  IObservationPosition,
   IPlaybackObservation,
   IPlaybackObserverEventType,
   IReadOnlyPlaybackObserver,
+  ObservationPosition,
   IFreezingStatus,
   IRebufferingStatus,
 } from "./playback_observer";
